requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
distro==1.9.0
elevenlabs==1.13.4
ffmpeg==1.4
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
MarkupSafe==3.0.2
moviepy==2.1.1
numpy==2.2.0
openai==1.57.3
pillow==10.4.0
proglog==0.1.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
Werkzeug==3.1.3
